1.暫時關閉要Reserve(自動OK) 2.Idle接著完成命令 3.Move合併TR部分

+++ Mirle.Agv/Model/MoveModel/Sr2000/BarcodeData.cs
@@ -18,7 +18,7 @@ namespace Mirle.Agv.Model
         public BarcodeData(int id, double x, double y)
         {
             ID = id;
-            ViewPosition = new MapPosition(x, y);
+            ViewPosition = new MapPosition((float)x, (float)y);
         }
     }
 }
